+++ packages/dash-table/src/dash-table/derived/cell/contents.tsx
@@ -5,12 +5,12 @@ import {
     ActiveCell,
     Data,
     Datum,
-    IVisibleColumn,
-    VisibleColumns,
+    DropdownValues,
     ICellFactoryProps,
     IViewportOffset,
-    DropdownValues,
-    Presentation
+    IVisibleColumn,
+    Presentation,
+    VisibleColumns
 } from 'dash-table/components/Table/props';
 import CellInput from 'dash-table/components/CellInput';
 import derivedCellEventHandlerProps, { Handler } from 'dash-table/derived/cell/eventHandlerProps';
@@ -19,6 +19,7 @@ import isCellEditable from './isEditable';
 import CellLabel from 'dash-table/components/CellLabel';
 import CellDropdown from 'dash-table/components/CellDropdown';
 import { memoizeOne } from 'core/memoizer';
+import getFormatter from 'dash-table/type/formatter';
 
 const mapData = R.addIndex<Datum, JSX.Element[]>(R.map);
 const mapRow = R.addIndex<IVisibleColumn, JSX.Element>(R.map);
@@ -63,58 +64,62 @@ class Contents {
         editable: boolean,
         isFocused: boolean,
         dropdowns: (DropdownValues | undefined)[][]
-    ): JSX.Element[][] => mapData(
-        (datum, rowIndex) => mapRow(
-            (column, columnIndex) => {
-                const active = isActiveCell(activeCell, rowIndex + offset.rows, columnIndex + offset.columns);
+    ): JSX.Element[][] => {
+        const formatters = R.map(getFormatter, columns);
 
-                const dropdown = dropdowns[rowIndex][columnIndex];
+        return mapData(
+            (datum, rowIndex) => mapRow(
+                (column, columnIndex) => {
+                    const active = isActiveCell(activeCell, rowIndex + offset.rows, columnIndex + offset.columns);
 
-                const isEditable = isCellEditable(editable, column.editable);
+                    const dropdown = dropdowns[rowIndex][columnIndex];
 
-                const className = [
-                    ...(active ? ['input-active'] : []),
-                    isFocused ? 'focused' : 'unfocused',
-                    'dash-cell-value'
-                ].join(' ');
+                    const isEditable = isCellEditable(editable, column.editable);
 
-                switch (getCellType(active, isEditable, dropdown, column.presentation)) {
-                    case CellType.Dropdown:
-                        return (<CellDropdown
-                            key={`column-${columnIndex}`}
-                            active={active}
-                            clearable={column.clearable}
-                            dropdown={dropdown}
-                            onChange={this.handlers(Handler.Change, rowIndex, columnIndex)}
-                            value={datum[column.id]}
-                        />);
-                    case CellType.Input:
-                        return (<CellInput
-                            key={`column-${columnIndex}`}
-                            active={active}
-                            className={className}
-                            focused={isFocused}
-                            onChange={this.handlers(Handler.Change, rowIndex, columnIndex)}
-                            onClick={this.handlers(Handler.Click, rowIndex, columnIndex)}
-                            onDoubleClick={this.handlers(Handler.DoubleClick, rowIndex, columnIndex)}
-                            onMouseUp={this.handlers(Handler.MouseUp, rowIndex, columnIndex)}
-                            onPaste={this.handlers(Handler.Paste, rowIndex, columnIndex)}
-                            type={column.type}
-                            value={datum[column.id]}
-                        />);
-                    case CellType.Label:
-                    default:
-                        return (<CellLabel
-                            className={className}
-                            key={`column-${columnIndex}`}
-                            onClick={this.handlers(Handler.Click, rowIndex, columnIndex)}
-                            onDoubleClick={this.handlers(Handler.DoubleClick, rowIndex, columnIndex)}
-                            value={datum[column.id]}
-                        />);
-                }
-            },
-            columns
-        ),
-        data
-    ));
+                    const className = [
+                        ...(active ? ['input-active'] : []),
+                        isFocused ? 'focused' : 'unfocused',
+                        'dash-cell-value'
+                    ].join(' ');
+
+                    switch (getCellType(active, isEditable, dropdown, column.presentation)) {
+                        case CellType.Dropdown:
+                            return (<CellDropdown
+                                key={`column-${columnIndex}`}
+                                active={active}
+                                clearable={column.clearable}
+                                dropdown={dropdown}
+                                onChange={this.handlers(Handler.Change, rowIndex, columnIndex)}
+                                value={datum[column.id]}
+                            />);
+                        case CellType.Input:
+                            return (<CellInput
+                                key={`column-${columnIndex}`}
+                                active={active}
+                                className={className}
+                                focused={isFocused}
+                                onChange={this.handlers(Handler.Change, rowIndex, columnIndex)}
+                                onClick={this.handlers(Handler.Click, rowIndex, columnIndex)}
+                                onDoubleClick={this.handlers(Handler.DoubleClick, rowIndex, columnIndex)}
+                                onMouseUp={this.handlers(Handler.MouseUp, rowIndex, columnIndex)}
+                                onPaste={this.handlers(Handler.Paste, rowIndex, columnIndex)}
+                                type={column.type}
+                                value={datum[column.id]}
+                            />);
+                        case CellType.Label:
+                        default:
+                            return (<CellLabel
+                                className={className}
+                                key={`column-${columnIndex}`}
+                                onClick={this.handlers(Handler.Click, rowIndex, columnIndex)}
+                                onDoubleClick={this.handlers(Handler.DoubleClick, rowIndex, columnIndex)}
+                                value={formatters[columnIndex](datum[column.id])}
+                            />);
+                    }
+                },
+                columns
+            ),
+            data
+        );
+    });
 }