+++ packages/dash-table/src/dash-table/index.ts
@@ -1,7 +1,7 @@
 import Environment from 'core/environment';
 import Logger from 'core/Logger';
 
-import DataTable from 'dash-table/DataTable';
+import DataTable from 'dash-table/dash/DataTable';
 
 Logger.setDebugLevel(Environment.debugLevel);
 Logger.setLogLevel(Environment.logLevel);